bot/requirements.txt
discord.py[voice]==1.2.2
websockets==8.0.1
youtube-dl==2019.7.30
flake8==3.7.7
mypy==0.720
pytest==5.0.1
pytest-asyncio==0.10.0
Sphinx==2.1.2
sphinx-autodoc-typehints==1.7.0
m2r==0.2.1
